+++ ViewModel/Adminhtml/Support/Index.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace Algolia\AlgoliaSearch\ViewModel\Adminhtml\Support;
+
+use Algolia\AlgoliaSearch\Helper\SupportHelper;
+use Algolia\AlgoliaSearch\ViewModel\Adminhtml\BackendView;
+use Magento\Backend\Block\Template;
+use Magento\Framework\View\Element\Block\ArgumentInterface;
+
+class Index implements ArgumentInterface
+{
+    /** @var BackendView */
+    private $backendView;
+
+    /** @var SupportHelper */
+    private $supportHelper;
+
+    /**
+     * @param BackendView $backendView
+     * @param SupportHelper $supportHelper
+     */
+    public function __construct(BackendView $backendView, SupportHelper $supportHelper)
+    {
+        $this->backendView = $backendView;
+        $this->supportHelper = $supportHelper;
+    }
+
+    /** @return bool */
+    public function isExtensionSupportEnabled()
+    {
+        return $this->supportHelper->isExtensionSupportEnabled();
+    }
+
+    /** @return string */
+    public function getApplicationId()
+    {
+        return $this->supportHelper->getApplicationId();
+    }
+
+    /**
+     * @return string
+     */
+    public function getLegacyVersionHtml()
+    {
+        /** @var Template $block */
+        $block = $this->backendView->getLayout()->createBlock(Template::class);
+
+        $block->setTemplate('Algolia_AlgoliaSearch::support/components/legacy-version.phtml');
+        $block->setData('extension_version', $this->supportHelper->getExtensionVersion());
+
+        return $block->toHtml();
+    }
+}
