+++ view/adminhtml/layout/algolia_algoliasearch_support_contact.xml
@@ -7,8 +7,12 @@
     </head>
     <body>
         <referenceContainer name="content">
-            <block class="Algolia\AlgoliaSearch\Block\Adminhtml\Support\Contact"
-                   name="support_contact" template="support/contact.phtml" />
+            <block class="Magento\Backend\Block\Template"
+                   name="support_contact" template="Algolia_AlgoliaSearch::support/contact.phtml">
+                <arguments>
+                    <argument name="view_model" xsi:type="object">Algolia\AlgoliaSearch\ViewModel\Adminhtml\Support\Contact</argument>
+                </arguments>
+            </block>
         </referenceContainer>
     </body>
 </page>
